// src/anythingllm-policy/domain/anythingllm-operation.enum.ts
/**
 * Enum of AnythingLLM operations
 * Used for authorization decisions
 */
export enum AnythingLLMOperation {
  THREAD_CHAT = 'THREAD_CHAT',
  THREAD_CREATE = 'THREAD_CREATE',
  THREAD_LIST = 'THREAD_LIST',
  THREAD_HISTORY = 'THREAD_HISTORY',
  THREAD_EXPORT = 'THREAD_EXPORT',
  THREAD_ADMIN_VIEW = 'THREAD_ADMIN_VIEW',
  VECTOR_SEARCH = 'VECTOR_SEARCH',
  DOCUMENT_UPLOAD = 'DOCUMENT_UPLOAD',
  WORKSPACE_CREATE = 'WORKSPACE_CREATE',
  WORKSPACE_LIST = 'WORKSPACE_LIST',
  SYSTEM_AUTH_CHECK = 'SYSTEM_AUTH_CHECK',
  SYSTEM_READ = 'SYSTEM_READ',
  VECTOR_COUNT_READ = 'VECTOR_COUNT_READ',
  WORKSPACE_COUNT_READ = 'WORKSPACE_COUNT_READ',
  DOCUMENT_COUNT_READ = 'DOCUMENT_COUNT_READ',
}